requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
filelock==3.13.1
geckodriver-autoinstaller==0.1.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
jsoncodable==0.1.7
jsonpickle==3.0.2
k-selenium-cookies==0.0.4
kproxy==0.0.1
noraise==0.0.16
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
PySocks==1.7.1
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
requests-file==1.5.1
selenium==4.16.0
selenium-browser==0.0.15
selenium-chrome==0.0.29
selenium-chromedriver==1.3
selenium-firefox==2.0.8
selenium-webdriver-extender==0.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
start-selenium-webdriver==0.3.1
tldextract==5.1.1
trio==0.23.1
trio-websocket==0.11.1
undetected-chromedriver==3.5.4
urllib3==2.1.0
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
xpath-utils==0.0.3
